+++ Breathy/source/src/main/java/com/apps/philipps/source/BreathInterpreter.java
@@ -91,7 +91,7 @@ public abstract class BreathInterpreter {
         boolean readyToAdd = false;
         float founds = 0;
         for (int i = 0; i < data.length - 2 && data[i] != null && data[i + 1] != null; i++) {
-            float d = data[i].data;
+            double d = data[i].data;
             if (founds < 2) {
                 if (d - norm > 0) {
                     if (moment == BreathMoment.None)
